Fix an issue with channel retriggers after muting. Pulse 2 works!
Turns out writing $4015 clears the length counter, and despite it
being halted, if the value is currently 0 the channel is muted. To
correct this you're meant to reload the counter with a write to
either $4003/$4007, but we were dutifully skipping that if our
shadow register hadn't changed. We need to *always* make this
write at the start of a note trigger, otherwise the new note
stays muted.