Reset global duration when performing a jump (Bxx) command
Turns out FamiTracker expects all frames to begin with global
duration disabled. Ordinarily the last note in any compiled pattern
takes care of turning it off, but if we skip that with Bxx before
that gets a chance to play, things get messy *real* quick. Fixes
the yakra demo, finally