Include our own copy of longbranch, rather than relying on ca65's
In testing I've found support for `.macpack` to be iffy at best
on Windows systems, and it's always a pain to try to debug via
someone who's never seen the error before. It's not a small enough
file to justify not including as part of the library, so let's
do that.