Implement a fake 50% square with no whine!
This is *coooomplicated,* but the idea is to emulate sortof a
50% square on top of a triangle, by alternating a slope instead
of trying to hold the thing actually flat. It sounds rather
convincing at louder square volumes, and it ditches the ugly $55
whine, but the drawback is that it doesn't really do "quiet."
Instead of "quiet" you get "more like a triangle", so for actual
compositions you might wish to switch to a different timbre, if
you need softer sounds.