Tweak mixing levels, fix a triangle transcription error
Less of an intentional error, and more an artifact of a previous
version that had a triangle bassline on DPCM, which sounded bad
played against the hardware triangle. Now that the bassline is
a sawtooth, this is much less bad sounding and the harmony skipping
a note is much more of a problem.