Fix demo effect conversion bug again
Issue #9 again, but in the opposite direction. Lxx is now loaded as N163
Zxx in Dn-FT 0.5.0.2 beta.

In issue #9, I erroneously determined that effect Lxx should be saved as
0x24 instead of 0x21. Through recent further testing, Dn-FT 0.4.0.1, a
version before the bug was introduced, saved Lxx as bytecode 0x21,
meaning that this module was re-saved with the wrong effect order.

Additionally, Dn-FT 0.5.0.1 has not entirely fixed the effects
conversion, saving the bytecode correctly but loading the bytecode
incorrectly due to a malformed module header, causing Dn-FT to interpret
it as a FT 0.5.0 beta module, instead of a Dn-FT module.
A fix has already been merged into the main branch.